AMX: Change AM Medium/Longs nbytes arg to size_t, instead of int
AM-2 specifies the nbytes argument to AM{Request,Reply}{I,Xfer} have
type `int`. This was an ILP32-centric decision that is now obsolete,
especially for the case of Longs where the payload could potentially
exceed 2GB in size (although AM_MaxLong for AM{UDP,MPI} currently
prevent this in practice). This change (widening the input argument
to injection) should have no impact on current clients.

More importantly, AM-2 specifies that AM Medium/Long handlers take their
nbytes argument as type `int`. On LP64 this is fundamentally incompatible
with GASNet AM Medium/Long handlers that take their nbytes argument as
`size_t`, leading to a language compliance violation when AMX called
GASNet-registered AM Medium/Long handlers.  It's surprising this
was working at all on {udp,mpi}-conduits.  An earlier commit that
fixed the AMX handler dispatch macros also applied this change
(which DOES have an impact on AM handler declarations in portable AM-2
clients).  This commit documents this change and fixes handlers in
our AM-2 test code to comply.

The nbytes type adjustment is propagated through injection to prevent
confusion, although the on-the-wire protocol remains unchanged
(currently a 16-bit unsigned field in the packet of each library).

+++ other/ammpi/ammpi.h
@@ -383,40 +383,40 @@ extern int AM_Poll(eb_t bundle);
 extern int AMMPI_Request(ep_t request_endpoint, ammpi_node_t reply_endpoint, handler_t handler, 
                          int numargs, ...);
 extern int AMMPI_RequestI (ep_t request_endpoint, ammpi_node_t reply_endpoint, handler_t handler, 
-                          void *source_addr, int nbytes,
+                          void *source_addr, size_t nbytes,
                           int numargs, ...);
 extern int AMMPI_RequestXfer(ep_t request_endpoint, ammpi_node_t reply_endpoint, handler_t handler, 
-                          void *source_addr, int nbytes, uintptr_t dest_offset, 
+                          void *source_addr, size_t nbytes, uintptr_t dest_offset, 
                           int async,
                           int numargs, ...);
 
 extern int AMMPI_Reply(void *token, handler_t handler, 
                          int numargs, ...);
 extern int AMMPI_ReplyI(void *token, handler_t handler, 
-                          void *source_addr, int nbytes,
+                          void *source_addr, size_t nbytes,
                           int numargs, ...);
 extern int AMMPI_ReplyXfer(void *token, handler_t handler, 
-                          void *source_addr, int nbytes, uintptr_t dest_offset, 
+                          void *source_addr, size_t nbytes, uintptr_t dest_offset, 
                           int numargs, ...);
 
 /* alternate forms that take va_list ptr to support GASNet */
 extern int AMMPI_RequestVA(ep_t request_endpoint, ammpi_node_t reply_endpoint, handler_t handler, 
                          int numargs, va_list argptr);
 extern int AMMPI_RequestIVA(ep_t request_endpoint, ammpi_node_t reply_endpoint, handler_t handler, 
-                          void *source_addr, int nbytes,
+                          void *source_addr, size_t nbytes,
                           int numargs, va_list argptr);
 extern int AMMPI_RequestXferVA(ep_t request_endpoint, ammpi_node_t reply_endpoint, handler_t handler, 
-                          void *source_addr, int nbytes, uintptr_t dest_offset, 
+                          void *source_addr, size_t nbytes, uintptr_t dest_offset, 
                           int async,
                           int numargs, va_list argptr);
 
 extern int AMMPI_ReplyVA(void *token, handler_t handler, 
                          int numargs, va_list argptr);
 extern int AMMPI_ReplyIVA(void *token, handler_t handler, 
-                          void *source_addr, int nbytes,
+                          void *source_addr, size_t nbytes,
                           int numargs, va_list argptr);
 extern int AMMPI_ReplyXferVA(void *token, handler_t handler, 
-                          void *source_addr, int nbytes, uintptr_t dest_offset, 
+                          void *source_addr, size_t nbytes, uintptr_t dest_offset, 
                           int numargs, va_list argptr);
